refactor homework: models and admin
