+++ src/homework/admin/__init__.py
@@ -1,0 +1,2 @@
+import homework.admin.question.admin
+import homework.admin.answer.admin
