+++ src/homework/querysets/__init__.py
@@ -1,0 +1,6 @@
+from homework.querysets.answer_query_set import AnswerQuerySet
+from homework.querysets.answer_access_log_entry_query_set import AnswerAccessLogEntryQuerySet
+__all__ = [
+    'AnswerQuerySet',
+    'AnswerAccessLogEntryQuerySet',
+]
